refactor: extract license reference related logic from ScanCodeDataProvider
